Test: added unit test for chained sessions with keyshare server
