+++ internal/sessiontest/helper_dosession_test.go
@@ -2,6 +2,8 @@ package sessiontest
 
 import (
 	"encoding/json"
+	"fmt"
+	"github.com/go-errors/errors"
 	"io/ioutil"
 	"path/filepath"
 	"testing"
@@ -301,4 +303,38 @@ func doSession(
 	}
 
 	return &requestorSessionResult{serverResult, clientResult, nil, dismisser}
+}
+
+func doChainedSessions(t *testing.T, conf interface{}, id irma.AttributeTypeIdentifier, opts ...option) {
+	client, handler := parseStorage(t, opts...)
+	defer test.ClearTestStorage(t, client, handler.storage)
+
+	require.IsType(t, IrmaServerConfiguration, conf)
+	irmaServer := StartIrmaServer(t, conf.(func() *server.Configuration)())
+	defer irmaServer.Stop()
+	nextServer := StartNextRequestServer(t, &irmaServer.conf.JwtRSAPrivateKey.PublicKey, id)
+	defer func() {
+		_ = nextServer.Close()
+	}()
+
+	var request irma.ServiceProviderRequest
+	require.NoError(t, irma.NewHTTPTransport(nextSessionServerURL, false).Get("1", &request))
+
+	// In case of chained sessions, the server's session store is queried twice in a single
+	// HTTP handler (when processing the irmaclient's response). The mutexes involved have caused
+	// deadlocks in the past when the frontend polls the session status, so we simulate polling in
+	// this test.
+	doSession(t, &request, client, irmaServer, nil, nil, nil, append(opts, optionPolling)...)
+
+	// check that our credential instance is new
+	issuedCredID := request.SessionRequest().Disclosure().Disclose[0][0][0].Type.CredentialTypeIdentifier()
+
+	for _, cred := range client.CredentialInfoList() {
+		if issuedCredID.String() == fmt.Sprintf("%s.%s.%s", cred.SchemeManagerID, cred.IssuerID, cred.ID) &&
+			cred.SignedOn.After(irma.Timestamp(time.Now().Add(-1*irma.ExpiryFactor*time.Second))) {
+			return
+		}
+	}
+
+	require.NoError(t, errors.New("newly issued credential not found in client"))
 }
